fsck: improve readahead
if a buffer isn't filled with directory entries,
this commit makes fsck not read blocks ahead of time.
and if fsck reads the last, this commit makes
fsck read blocks of the next direcotry in advance.

64GB Samsung Class10 UHS-I Pro sdcard is filled
with 9947 directories and 16504 files by fsstress.

After this patch, the elpased time decreases
from 24.95 secs to 11.61 secs by about 46%.

Signed-off-by: Hyunchul Lee <hyc.lee@gmail.com>