mkfs: place bitmap immediately before alignment boundary if possible
Prior to this commit, when the boundary alignment is set to the allocation
unit of flash media, then the FAT and the bitmap end up in different AUs.
This can have a deleterious effect on performance and endurance of some
flash media. This commit adds an opportunistic attempt to move the bitmap
to just before the alignment boundary (instead of just after it) to allow
the FAT and the bitmap to share an allocation unit where possible.

Signed-off-by: Matt Whitlock <exfatprogs@mattwhitlock.name>